+++ 1-Elementary-Algorithms/1.2.9_Knapsack_(Unbounded_and_Coin_Change).cpp
@@ -34,7 +34,7 @@ Space Complexity:
 long long unbounded_knapsack(
     const std::vector<int> &weight, const std::vector<long long> &value, int capacity
 ) {
-  int n = weight.size();
+  int n = static_cast<int>(weight.size());
   std::vector<long long> dp(capacity + 1, 0);
   for (int i = 0; i < n; i++) {
     for (int w = weight[i]; w <= capacity; w++) {
