+++ 3-Strings/3.5.3_Sequence_Alignment.cpp
@@ -37,7 +37,7 @@ using std::string;
 std::pair<string, string> align_sequences(
     const string &s1, const string &s2, int gap_cost = 1, int sub_cost = 1
 ) {
-  int n = s1.size(), m = s2.size();
+  int n = static_cast<int>(s1.size()), m = static_cast<int>(s2.size());
   std::vector<std::vector<int>> dp(n + 1, std::vector<int>(m + 1, 0));
   for (int i = 0; i <= n; i++) {
     dp[i][0] = i * gap_cost;
