+++ 3-Strings/3.6.1_Suffix_Array_and_LCP_(Manber-Myers).cpp
@@ -47,7 +47,7 @@ class SuffixArrayManberMyers {
 
  public:
   explicit SuffixArrayManberMyers(const string &s) : s(s), sa(s.size()), rk(s.size()) {
-    int n = s.size();
+    int n = static_cast<int>(s.size());
     std::iota(sa.begin(), sa.end(), 0);
     for (int i = 0; i < n; i++) {
       rk[i] = static_cast<unsigned char>(s[i]);
@@ -67,7 +67,7 @@ class SuffixArrayManberMyers {
   const std::vector<int> &get_sa() const { return sa; }
 
   std::vector<int> get_lcp() const {
-    int n = s.size();
+    int n = static_cast<int>(s.size());
     if (n == 0) {
       return {};  // Avoid constructing a vector of size (size_t)(-1).
     }
