+++ 6-Mathematics/6.6.2_Fast_Fourier_Transform.cpp
@@ -37,7 +37,7 @@ typedef std::complex<double> cd;
 const double PI = acos(-1.0);
 
 void fft(std::vector<cd> &a, bool invert) {
-  int n = a.size();
+  int n = static_cast<int>(a.size());
   for (int i = 1, j = 0; i < n; i++) {
     int bit = n >> 1;
     for (; j & bit; bit >>= 1) {
